Solar Karborundum Migration
Current heating systems mean that previous harvesting altitudes simply can't be reached.  I'm not married to these numbers, but they push that harvestable band out further - centered on 150Mm, extending to around 200Mm.  This allows for careful ship construction to succeed but still is close enough to be hazardous, with incentives for going deeper in to mine the richer deposits if the ship can handle it.